+++ src/mongoclasses/sync.py
@@ -1,8 +1,13 @@
 from dataclasses import asdict
-from .dataclass import fromdict, create_include_dict_factory
+from .dataclass import (
+    fromdict, create_include_dict_factory, _is_mongoclass_instance, _is_mongoclass_type
+)
 
 
 def insert_one(obj, /):
+    if not _is_mongoclass_instance(obj):
+        raise TypeError("Not a mongoclass instance.")
+
     document = asdict(obj)
     if document["_id"] is None:
         del document["_id"]
@@ -13,6 +18,9 @@ def insert_one(obj, /):
 
 
 def update_one(obj, /, fields=None):
+    if not _is_mongoclass_instance(obj):
+        raise TypeError("Not a mongoclass instance.")
+
     dict_factory = dict if fields is None else create_include_dict_factory(fields)
     document = asdict(obj, dict_factory=dict_factory)
     return type(obj).collection.update_one(
@@ -21,6 +29,9 @@ def update_one(obj, /, fields=None):
 
 
 def delete_one(obj, /):
+    if not _is_mongoclass_instance(obj):
+        raise TypeError("Not a mongoclass instance.")
+
     return type(obj).collection.delete_one({"_id": obj._id})
 
 
@@ -28,6 +39,9 @@ def find_one(cls, /, query, fromdict=fromdict):
     """
     Return a single instance that matches the query on the mongoclass or None.
     """
+    if not _is_mongoclass_type(cls):
+        raise TypeError("Not a mongoclass type.")
+
     document = cls.collection.find_one(query)
     if document is None:
         return None
@@ -39,4 +53,7 @@ def find(cls, /, query):
     Performs a query on the mongoclass.
     Returns a DocumentCursor.
     """
+    if not _is_mongoclass_type(cls):
+        raise TypeError("Not a mongoclass type.")
+
     return cls.collection.find(filter=query)
